backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<4.1
pydantic-settings==2.7.1
email-validator==2.2.0
